+++ Exodus/src/utopia/exodus/rest/resource/ExodusResources.scala
@@ -3,7 +3,7 @@ package utopia.exodus.rest.resource
 import utopia.exodus.rest.resource.description.{DescriptionRolesNode, LanguageFamiliaritiesNode, LanguagesNode, RolesNode, TasksNode}
 import utopia.exodus.rest.resource.device.DevicesNode
 import utopia.exodus.rest.resource.organization.OrganizationsNode
-import utopia.exodus.rest.resource.user.UsersNode
+import utopia.exodus.rest.resource.user.{QuestSessionsNode, UsersNode}
 
 /**
   * Lists all resources introduced in the Exodus project
@@ -25,7 +25,7 @@ object ExodusResources
 	/**
 	  * All resources that require authorization for at least some features
 	  */
-	val authorized = Vector(DevicesNode, OrganizationsNode, RolesNode, TasksNode)
+	val authorized = Vector(DevicesNode, OrganizationsNode, RolesNode, TasksNode, QuestSessionsNode)
 	
 	
 	// COMPUTED	---------------------------------